Refactor: Isolate database initialization and fix imports
