+++ blog/fault_tolerance_perfect.py
@@ -31,8 +31,8 @@ from sqlalchemy.pool import QueuePool
 # import kubernetes  # Необязательный импорт
 from prometheus_client import Counter, Histogram, Gauge, start_http_server
 
-from blog import db
-from blog import db as database
+from blog.database import db
+from blog.database import db as database
 
 class HealthStatus(Enum):
     """Статусы здоровья"""
