+++ blog/routes/ai_admin.py
@@ -6,7 +6,7 @@ from flask import Blueprint, render_template, request, redirect, url_for, flash,
 from flask_login import login_required, current_user
 from blog.models import Post, Category, User, Comment
 from blog.ai_content_perfect import PerfectAIContentGenerator, ContentScheduler, populate_blog_with_ai_content
-from blog import db
+from blog.database import db
 from functools import wraps
 import threading
 import os
