+++ blog/routes/auth.py
@@ -6,7 +6,7 @@ from flask import Blueprint, render_template, redirect, url_for, flash, request
 from flask_login import login_user, logout_user, login_required, current_user
 from blog.models import User
 from blog.forms import LoginForm, RegisterForm, ProfileForm
-from blog import db
+from blog.database import db
 
 bp = Blueprint('auth', __name__)
