+++ blog/routes/seo.py
@@ -8,7 +8,7 @@ from blog.models import Post, Category, Tag, User
 from blog.advanced_seo import advanced_seo_optimizer
 from blog.auto_seo_optimizer import AutoSEOOptimizer
 from blog.seo_analytics import SEOAnalytics
-from blog import db
+from blog.database import db
 from datetime import datetime
 import json
