+++ blog/ui_perfect.py
@@ -17,8 +17,8 @@ import redis
 from werkzeug.contrib.cache import SimpleCache
 
 from blog.models import Post, User, Category, Comment, Tag
-from blog import db
-from blog import db as database
+from blog.database import db
+from blog.database import db as database
 
 class ThemeType(Enum):
     """Типы тем"""
