+++ scripts/migrate_database_imports.py
@@ -147,11 +147,15 @@ def main():
     for file in all_files_to_update:
         print(f"  - {file}")
     
-    # Запрашиваем подтверждение
-    response = input("\n❓ Продолжить миграцию? (y/N): ")
-    if response.lower() != 'y':
-        print("❌ Миграция отменена")
-        return
+    # Автоматическое подтверждение для CI/CD
+    import sys
+    if '--auto' in sys.argv:
+        print("\n✅ Автоматический режим активирован")
+    else:
+        response = input("\n❓ Продолжить миграцию? (y/N): ")
+        if response.lower() != 'y':
+            print("❌ Миграция отменена")
+            return
     
     # Выполняем миграцию
     print("\n🔧 Выполнение миграции...")
